pyproject.toml
[project]
name = "haomnilogic-local"
version = "1.0.0b4"
description = "A Home Assistant integration for Hayward OmniLogic/OmniHub pool controllers using the local UDP api"
readme = "README.md"
requires-python = ">=3.14.2"
authors = [
    { name = "cryptk", email = "cryptk@users.noreply.github.com" },
    { name = "djtimca" },
    { name = "garionphx" },
]
license-files = ["LICENSE"]
dependencies = ["python-omnilogic-local"]

[project.urls]
repository = "https://github.com/cryptk/haomnilogic-local"


[dependency-groups]
dev = [
    "mypy>=1.18.2,<2.0.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.2.0,<2.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "prek>=0.3.9",
    "pytest-homeassistant-custom-component>=0.13.323",
]

[tool.codespell]
ignore-words-list = ["hass"]

[tool.mypy]
python_version = "3.14"
plugins = ["pydantic.mypy"]
follow_imports = "normal"
strict = true
# warn_return_any = false


[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = [
    "ERA",   # eradicate commented code
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    # "EM",     # flake8-errmsg
    # "FIX",    # flake8-fixme
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "LOG", # flake8-logging
    "G",   # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    # "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "TID252", # flake8-tidy-imports
    # "TD",     # flake8-todos
    # "TC",     # flake8-type-checking
    # "ARG",    # flake8-unused-arguments
    "I", # isort
    # "N",      # pep8-naming
    # "PERF",   # perflint
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    # "DOC",   # pydoclint, only available with ruff preview mode enabled
    # "D",      # pydocstyle
    "F", # pyflakes
    # "UP",     # pyupgrade
    # "RUF",    # ruff-specific rules
    # "TRY",    # tryceratops
]
ignore = [
    "D100", # Disabled until we get everything documented.
    "D101", # Disabled until we get everything documented.
    "D102", # Disabled until we get everything documented.
    "D107", # Disabled until we get everything documented.
]
future-annotations = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-import-conventions.extend-aliases]
voluptuous = "vol"
"homeassistant.helpers.area_registry" = "ar"
"homeassistant.helpers.config_validation" = "cv"
"homeassistant.helpers.device_registry" = "dr"
"homeassistant.helpers.entity_registry" = "er"
"homeassistant.helpers.issue_registry" = "ir"
